+++ lib/src/pages/levantamiento/first_step.dart
@@ -6,7 +6,7 @@ import 'package:geolocator/geolocator.dart';
 import 'package:image_picker/image_picker.dart';
 
 
-import 'package:movilidad/src/levantamiento/model/levantamiento.dart';
+import 'package:movilidad/src/pages/levantamiento/model/levantamiento.dart';
 
 class firstStep extends StatefulWidget {
   //objeto
