+++ lib/src/pages/levantamiento/model/afectado.dart
