+++ lib/src/pages/view/home_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 
-import 'package:movilidad/src/home_screens/home_screen1.dart';
-import 'package:movilidad/src/home_screens/home_screen2.dart';
-import 'package:movilidad/src/home_screens/home_screen3.dart';
+import 'package:movilidad/src/pages/home_screens/home_screen1.dart';
+import 'package:movilidad/src/pages/home_screens/home_screen2.dart';
+import 'package:movilidad/src/pages/home_screens/home_screen3.dart';
 
 
 class HomeScreen extends StatefulWidget {
