+++ lib/src/pages/view/login_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:movilidad/src/view/home_screen.dart';
+import 'package:movilidad/src/pages/view/home_screen.dart';
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({Key? key}) : super(key: key);
